clean-up, update DeltaFinder defaults
