+++ CalPatRec/test/deltaFinder_test0.fcl
@@ -19,32 +19,7 @@ physics : {
 
     producers : { @table::Reconstruction.producers
         DeltaFinder: { @table::CalPatRec.producers.DeltaFinder
-            diagLevel                     :  0  ## 0 ##
-            debugLevel                    :  0  ## 1 ##
             useTimePeaks                  :  0
-            maxChi2Seed                   : 16. ## 10 16 used by findSeeds, for each hit
-            maxChi2Radial                 : 12. ## 10 25 used by recoverStation (wire-to-seed dist)
-            maxChi2All                    : 12. ## 10 16 used by pruneSeeds to mark bad seeds
-            maxDxy                        : 30. ## used by connectSeeds 30 40 mm
-            maxGap                        :  3  ## 3 gap in missing stations along the candidate
-            sigmaR                        :  5. # 10 mm, used in the calculation of chi2Radial, roughly 3 MeV/c
-            seedRes                       :  5. # 5 is good just to see what it does
-            maxDriftTime                  : 35. ## 40 ns, includes some tolerance (50)
-            maxStrawDt                    : 10. # max sensible time difference for a straw, ns, off by default
-            maxDtDs                       :  5. # ns, max allowed T0 shift per station
-            maxDtDc                       : 10. ## 15 20 30 ns, max dT between two delta candidates
-            maxHitSeedDt                  : 10  ##
-            minDeltaNHits                 :  5
-
-            diagPlugin : { @table::CalPatRec.producers.DeltaFinder.diagPlugin
-                mcDiag                    : true
-                printElectrons            : 1 ## 0 ##
-                printElectronsHits        : 0 ## 1 ##
-                printElectronsMinMom      : 0.
-                printElectronsMaxMom      : 20.
-                printElectronsMinNHits    : 10   ## 30
-                printElectronsMaxFReco    : 0.2 ## 0.5
-            }
         }
 
         DeltaFinderAna: { @table::CalPatRec.producers.DeltaFinderAna
